+++ test/services/multichain/MultiChainStatusProcessor.test.ts
@@ -5,6 +5,8 @@ import {getTestContainer} from '../../helpers/getTestContainer';
 import {MultiChainStatusProcessor} from '../../../src/services/multiChain/MultiChainStatusProcessor';
 import {ChainStatusWithAddress} from '../../../src/types/ChainStatus';
 import {chainStatusWithAddressFactory, chainStatusFactory} from '../../mocks/factories/chainStatusFactory';
+import {loadTestEnv} from "../../helpers/loadTestEnv";
+import mongoose from "mongoose";
 
 describe('MultiChainStatusProcessor', () => {
   let chainStatusWithAddress: ChainStatusWithAddress[];
@@ -20,6 +22,11 @@ describe('MultiChainStatusProcessor', () => {
     },
   ];
 
+  before(async () => {
+    const config = loadTestEnv();
+    await mongoose.connect(config.MONGODB_URL, {useNewUrlParser: true, useUnifiedTopology: true});
+  });
+
   beforeEach(() => {
     const container = getTestContainer();
     container.bind(MultiChainStatusProcessor).toSelf();
@@ -29,8 +36,8 @@ describe('MultiChainStatusProcessor', () => {
   describe('when all chains can mint', () => {
     it('returns the chainsIdsReadyForBlock for all chains', async () => {
       const result = await multiChainStatusProcessor.apply(chainStatusWithAddress, 10);
-      expect(result.chainsStatuses[0]).to.includes({chainId: 'bsc'});
 
+      expect(result.chainsStatuses[0]).to.includes({chainId: 'bsc'});
       expect(result.chainsIdsReadyForBlock).to.deep.equal(['bsc', 'avax']);
     });
   });