+++ CHANGELOG.md
@@ -6,6 +6,17 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+
+## [7.7.0] - 2023-01-26
+### Added
+- cache masterchain data required to run consensus
+
+### Changed
+- set default timeout to 5s for `chain.getStatus()`
+
+### Removed
+- remove deprecated leader selection
+
 ## [7.6.6] - 2023-01-26
 - Fixed GitHub Action "CI"